requirements.txt
fastapi
uvicorn
faiss-cpu
python-dotenv
PyMuPDF
python-docx
requests
sentence-transformers
langchain